Removed `+` from sbt actions used in travis
Despite the build changes in the previous commit where projects have been forced
to use Scala 2.11, executing the following command:

    sbt +test plugin/scripted

would result in an error in the play-ebean-root project, because some
dependencies cannot be resolved (eg, play-java-jdbc_2.10;2.5.0-M1). What's odd
is that executing each command independently succeed. Furthermore, also swapping
the two command (ie., running: sbt plugin/scripted +test) succeeds.

Because the `+` is used for executing the same action for each cross scala
version, and considering that currently only Scala 2.11 is being used, I'm
removing the `+` so that we can make progress and announce the release of Play
2.5.0-M1.