Explicitly set `crossScalaVersions` to be empty for the `plugin` project
When running the command `sbt +plugin/scripted`, the build used to fail because
it would try to find play 2.5.0-M1 dependencies cross-compiled with Scala 2.10,
which of course do not exist. To work around the issue I've initially removed
the `+` from the sbt commands ran by travis, but that solution wasn't good
enough as the release script would still fail the build. Hence, I've reverted
the mentioned commit (see ), and instead I'm explicitly setting
`crossScalaVersions` to be empty for the `plugin` project.